Refactor Http classes to be data-generic and add file serving
Changed HttpRequest, HttpResponse and HttpMessage classes to generic classes to allow user-defined types for request data.

Added data field in HttpRequest class which can hold any type that meets the requirements of being default constructible. This provides a way to easily attach additional information to a request.

For HttpResponse class, the request field now holds a HttpRequest object that also is a template object.

Changes were made to HttpRequest and HttpResponse, including constructors and methods, to support this new template element.

Declare an EmptyRequestData struct as an example of user-defined request body data.

Added cmake-build-release/ to the .gitignore file for ignoring release build files.

The HTTP server now has the capability to serve files. It can be told which file extensions it can serve. For performance, server has the option to load and cache files at launch.

In interceptor classes, template support was also added for offering additional flexibility for incoming and outgoing methods which now receive the templated version of request and responses.

In HttpServer class added methods for setting allowed file extensions, loading and caching files, and getting a file response. The files variable needs to be true to enable serving files, default is false.

Updated the HttpParser to work with the newly templated HttpResponse and HttpRequest objects.

Updated example_Socket.cpp to reflect these changes and also added an example on how to use the newly added functionality with custom request data.